backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
numpy>=2.0.0
opencv-python-headless>=4.10.0
imagehash>=4.3.1
scipy>=1.13.0
scikit-image>=0.24.0
